ashmem: int/size_t and slight modernization.
ashmem_get_size_region() should return size_t, since
ashmem_create_region() takes a size_t.

This is a no-op (other than sign) for LP32, since size_t is 32 bits.

For LP64, this is an ABI change, but in effect it just means that we
don't zero the top 32 bits of the result register; any code that worked
before still works, and (highly unlikely) code that allocated > 2GiB
ashmem regions can now happily query that allocated size *provided*
that the implementation is memfd, not ashmem. The ashmem implementation
can only return an int from the kernel anyway, so it's already
truncated before it gets to this point :-(

Also lose bogus "\n"s from logging, and use unique_fd rather than goto.

Bug: http://b/205348471
Test: treehugger

Former-commit-id: 4f1a0c84ff5d897fc7c44ed1ec8e34faf6fd0b62
Former-commit-id: cd2bd695d23bfc4c0e5282b0b029e0bf6f7b5522
Former-commit-id: fa4a522b83fe7ed796ac5950480bb40b391ef82e

+++ libcutils/ashmem_test.cpp
@@ -33,7 +33,7 @@ void TestCreateRegion(size_t size, unique_fd &fd, int prot) {
     fd = unique_fd(ashmem_create_region(nullptr, size));
     ASSERT_TRUE(fd >= 0);
     ASSERT_TRUE(ashmem_valid(fd));
-    ASSERT_EQ(size, static_cast<size_t>(ashmem_get_size_region(fd)));
+    ASSERT_EQ(size, ashmem_get_size_region(fd));
     ASSERT_EQ(0, ashmem_set_prot_region(fd, prot));
 
     // We've been inconsistent historically about whether or not these file
